requirements.txt
paho-mqtt
aiocoap
pandas
matplotlib
ipykernel